transform wallet export dialog to separate tab
