+++ .env.example
@@ -9,6 +9,6 @@ STATS_PWD =admin
 CARDANOLITE_DEMO_WALLET_MNEMONIC = 'civil void tool perfect avocado sweet immense fluid arrow aerobic boil flash'
 
 # mock server config
-CARDANOLITE_ENABLE_SERVER_MOCKING_MODE = true
+CARDANOLITE_ENABLE_SERVER_MOCKING_MODE = false
 CARDANOLITE_MOCK_TX_SUBMISSION_SUCCESS = false
 CARDANOLITE_MOCK_TX_SUMMARY_SUCCESS = false
