+++ app/public/css/styles.css
@@ -60,6 +60,11 @@ h2 {
   padding-bottom: 0.25em;
   padding-left: 0.2rem;
 }
+@media screen and (max-width: 7px) {
+  h2 {
+    font-size: 1.5rem;
+  }
+}
 h3 {
   margin-bottom: 0.25em;
   font-weight: 300;
@@ -414,7 +419,7 @@ div.box.box-auto {
   flex-direction: column;
   background-color: white;
 }
-.box.center {
+.center {
   text-align: center;
 }
 .overlay .box .message {
@@ -826,14 +831,13 @@ div.box.box-auto {
   height: 20px;
   animation: loading-spin 1s linear infinite;
 }
-.loading-button {
+.submit-button {
   display: flex;
   align-items: center;
   font-size: 1.1rem;
   font-weight: 600;
   float: right;
   margin: 15px 5px;
-  margin-top: 2em;
   height: 2.6rem;
   width: 200px;
   text-align: center;
